+++ Packages/ai.fxn.fxn3d/Runtime/Services/Predictor.cs
@@ -183,7 +183,7 @@ namespace Function.Services {
         predictions
         created
         description
-        readme
+        card
         media
         acceleration
         signature {{